♻️ refactor: simplify user retrieval and deletion logic by improving ID handling

+++ src/api/admin/user.ts
@@ -1,13 +1,12 @@
 import { PrismaClient } from "@prisma/client";
 import express from "express";
-import { AuthRequest, authToken } from "../../middlewares/authMiddleware";
-import { isAdmin } from "../../middlewares/adminMiddleware";
+import { AuthRequest } from "../../middlewares/authMiddleware";
 import { asyncHandler } from "../../utils/asyncHandler";
 
 const prisma = new PrismaClient();
 const router = express.Router();
 
-router.get("/users", asyncHandler(async (req: AuthRequest, res) => {
+router.get("/users", asyncHandler(async (_, res) => {
   try {
     const users = await prisma.user.findMany({
       select: {
@@ -16,6 +15,9 @@ router.get("/users", asyncHandler(async (req: AuthRequest, res) => {
         nickname: true,
         role: true,
         createdAt: true,
+      },
+      orderBy: {
+        createdAt: "desc",
       },
     });
     res.status(200).json(users);
@@ -26,17 +28,26 @@ router.get("/users", asyncHandler(async (req: AuthRequest, res) => {
 
 router.delete("/users/:id", asyncHandler(async (req: AuthRequest, res) => {
   const { id } = req.params;
+  const userId = Number(id)
+
+  if(isNaN(userId)) {
+    return res.status(400).json({ message: "Invalid user ID" });
+  }
+
   try {
     const user = await prisma.user.findUnique({
-      where: { id: Number(id) },
+      where: { id: userId },
       select: { id: true, nickname: true, email: true },
     });
+
     if (!user) {
       return res.status(404).json({ message: "User not found" });
     }
+
     await prisma.user.delete({
-      where: { id: Number(id) },
+      where: { id: userId },
     });
+
     return res.status(200).json({ message: "User deleted successfully", user });
   } catch (err) {
     return res.status(500).json({ message: "Internal server error" });
